fix mistake 2

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/payment/controller/PaymentClient.java
@@ -4,6 +4,9 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import ru.yandex.practicum.payment.Dto.OrderDto;
 import ru.yandex.practicum.payment.Dto.PaymentDto;
+import ru.yandex.practicum.payment.enums.OrderState;
+
+import java.math.BigDecimal;
 
 @FeignClient(name = "payment", path = "/api/v1/payment")
 public interface PaymentClient {
@@ -12,7 +15,7 @@ public interface PaymentClient {
     PaymentDto createPayment(OrderDto orderDto);
 
     @PostMapping("/totalCost")
-    double totalCost(OrderDto orderDto);
+    BigDecimal totalCost(OrderDto orderDto);
 
     @PostMapping("/refund")
     void paymendRefund(String paymendId);
