+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -3,10 +3,12 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.payment.Dto.OrderDto;
 import ru.yandex.practicum.payment.Dto.PaymentDto;
 
+import java.math.BigDecimal;
+
 public interface PaymentService {
     PaymentDto createPayment(OrderDto orderDto);
 
-    double totalCost(OrderDto orderDto);
+    BigDecimal totalCost(OrderDto orderDto);
 
     void paymendRefund(String paymendId);
